phasmo_map_picker/requirements.txt
discord.py==2.5.2
dotenv==0.9.9
requests==2.32.3